# openapi.yaml
openapi: 3.1.0
info:
  title: PixRail Payment Switch API
  version: 1.0.0
  summary: Versioned API for Pix transfer intake, decisioning, SPI callbacks, and outbox inspection.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://api.pixrail.dev
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
  /healthz:
    get:
      operationId: healthCheck
      security: []
      summary: Health check
      responses:
        "200":
          description: Process is alive.
        "400":
          description: Malformed request.
  /readyz:
    get:
      operationId: readinessCheck
      security: []
      summary: Readiness check
      responses:
        "200":
          description: Runtime dependencies are available.
        "400":
          description: Malformed request.
  /metrics:
    get:
      operationId: prometheusMetrics
      security: []
      summary: Prometheus metrics
      responses:
        "200":
          description: Prometheus text exposition.
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Malformed request.
  /v1/pix/transfers:
    post:
      summary: Create an idempotent Pix transfer
      operationId: createPixTransfer
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
        - name: X-Correlation-ID
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransferRequest"
            examples:
              approved:
                value:
                  account_id: acct_123
                  amount_cents: 12345
                  currency: BRL
                  receiver_key: receiver@example.com
                  receiver_key_type: EMAIL
                  description: invoice 123
      responses:
        "201":
          description: Transfer accepted and newly created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResponse"
        "200":
          description: Idempotent replay of an existing transfer.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          $ref: "#/components/responses/DependencyFailed"
  /v1/pix/transfers/{id}:
    get:
      summary: Fetch a tenant-scoped Pix transfer
      operationId: getPixTransfer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transfer found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/pix/transfers/{id}/spi-submissions:
    post:
      summary: Submit a persisted accepted transfer to the simulated SPI rail
      operationId: submitPixTransferToSPI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SPI submission recorded or replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/pix/transfers/{id}/reviews:
    post:
      summary: Record a manual fraud review decision
      operationId: recordPixTransferReview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecisionRequest"
      responses:
        "200":
          description: Review decision recorded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/pix/transfers/{id}/spi-callbacks:
    post:
      summary: Record a simulated SPI settlement callback
      operationId: recordSPICallback
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettlementCallbackRequest"
      responses:
        "200":
          description: Callback recorded or replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/outbox:
    get:
      summary: Inspect tenant-authenticated outbox records for local operations
      operationId: listOutbox
      responses:
        "200":
          description: Outbox records.
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  responses:
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: API key is missing or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found or tenant-isolated.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: State conflict.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Token bucket exhausted.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    DependencyFailed:
      description: DICT or downstream dependency failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    CreateTransferRequest:
      type: object
      required: [account_id, amount_cents, currency, receiver_key, receiver_key_type]
      properties:
        account_id:
          type: string
        amount_cents:
          type: integer
          minimum: 1
          maximum: 2000000
        currency:
          type: string
          const: BRL
        receiver_key:
          type: string
        receiver_key_type:
          type: string
          enum: [CPF, CNPJ, EMAIL, PHONE, EVP]
        description:
          type: string
    SettlementCallbackRequest:
      type: object
      required: [spi_message_id, status, code]
      properties:
        spi_message_id:
          type: string
        status:
          type: string
          enum: [accepted, rejected]
        code:
          type: string
    ReviewDecisionRequest:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [approve, block]
        reason:
          type: string
    TransferResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/Transfer"
        meta:
          type: object
          required: [idempotent_replay]
          properties:
            idempotent_replay:
              type: boolean
    Transfer:
      type: object
      required: [id, tenant_id, account_id, status, amount_cents, currency, created_at, updated_at]
      properties:
        id:
          type: string
        tenant_id:
          type: string
        account_id:
          type: string
        status:
          type: string
          enum: [accepted, approved, blocked, review, settled, rejected]
        amount_cents:
          type: integer
        currency:
          type: string
        receiver_key_type:
          type: string
        receiver_name:
          type: string
        receiver_bank:
          type: string
        fraud_score:
          type: integer
        fraud_rules:
          type: array
          items:
            type: string
        decision_reason:
          type: string
        spi_message_id:
          type: string
        end_to_end_id:
          type: string
        settlement_code:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object